merging

+++ server/routes/orders.route.js
@@ -1,8 +1,9 @@
 const route = require('express').Router()
-const {getAll, create, deleteOne} = require("../controller/orders.controller")
+const {getAll, create, deleteOne, getByUserId} = require("../controller/orders.controller")
 
 
 route.get("/getAll",getAll)
+route.get("/getById",getByUserId)
 route.post("/createOrder",create)
 route.delete("/deleteOrder/:id",deleteOne)
